+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Added
+- `SUBSTR` function [MR186](https://github.com/Qrlew/qrlew/pull/186)
 
 ## [0.4.13] - 2023-11-14
 ## Added